make db exports from its own file to simplify importing models, and connecting db.
