+++ src/models/board.ts
@@ -1,15 +1,14 @@
-import Sequelize = require('sequelize');
-
-// const Board = Sequelize.define('board', {
-//   id: {
-//     type: Sequelize.UUID,
-//     defaultValue: Sequelize.UUIDV4,
-//     primaryKey: true,
-//   },
-//   name: Sequelize.STRING(255),
-//   timestamps: true,
-//   createdAt: 'date_created',
-//   updatedAt: 'date_updated',
-// });
-
-// module.exports = Board;
+// import Sequelize = require('sequelize');
+export const board = (sequelize, Sequelize) => {
+  return sequelize.define('board', {
+    id: {
+      type: Sequelize.UUID,
+      defaultValue: Sequelize.UUIDV4,
+      primaryKey: true,
+    },
+    name: Sequelize.STRING(255),
+    timestamps: true,
+    createdAt: 'date_created',
+    updatedAt: 'date_updated',
+  });
+};
